tests/wpt/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-in-cross-origin-ifame-002.sub.html

<!DOCTYPE html>
<head>
<title>An image with loading='lazy' in cross origin iframe loads when it gets
       visible by scrolling the parent scroll container of the iframe</title>
<link rel="help" href="https://github.com/scott-little/lazyload">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
</head>

<div style="height:1000vh;"></div>
<iframe id="iframe" width="500px" height="500px"></iframe>
<script>
promise_test(async t => {
  iframe.src =
    get_host_info().HTTP_NOTSAMESITE_ORIGIN +
    new URL("resources/", self.location).pathname +
    "image-loading-lazy-in-viewport-iframe.html";

  let image_loaded = false;

  await new Promise(resolve => {
    window.addEventListener("message", event => {
      if (event.data == "window_loaded") {
        resolve();
      } else if (event.data == "image_loaded") {
        image_loaded = true;
      }
    }, { once: true });
  });

  assert_false(image_loaded,
      "lazy-load image shouldn't block window load event");

  iframe.scrollIntoView();

  await new Promise(resolve => {
    window.addEventListener("message", event => {
      assert_equals(event.data, "image_loaded",
          "lazy-load image should be loaded once after it gets visible");
      resolve();
    });
  });
});
</script>
